refactor name and directory in rnn and lstm
