+++ zenu-autograd/src/nn/mod.rs
@@ -1,6 +1,5 @@
 pub mod batch_norm;
 pub mod conv2d;
 pub mod dropout;
-pub mod lstm;
 pub mod pool2d;
-pub mod rnn;
+pub mod rnns;
