+++ zenu-autograd/src/nn/rnns/rnn/cudnn.rs
@@ -6,36 +6,32 @@ use crate::{creator::alloc::alloc, Function, Variable, VariableWeak};
 
 use zenu_matrix::device::nvidia::Nvidia;
 
-struct CudnnLSTM<T: Num> {
+use super::RNNOutput;
+
+struct CudnnRNN<T: Num> {
     rnn_desc: Rc<RefCell<RNNDescriptor<T>>>,
     x: Variable<T, Nvidia>,
     weight: Variable<T, Nvidia>,
     hx: Option<Variable<T, Nvidia>>,
-    cx: Option<Variable<T, Nvidia>>,
     hy: VariableWeak<T, Nvidia>,
-    cy: VariableWeak<T, Nvidia>,
     y: VariableWeak<T, Nvidia>,
     is_training: bool,
 }
 
-impl<T: Num> Function<T, Nvidia> for CudnnLSTM<T> {
+impl<T: Num> Function<T, Nvidia> for CudnnRNN<T> {
     fn forward(&self) {
         let mut rnn_desc = self.rnn_desc.borrow_mut();
-        let hx = self.hx.as_ref().map(Variable::get_as_ref);
-        let cx = self.cx.as_ref().map(Variable::get_as_ref);
-        let output = rnn_desc.lstm_fwd(
+        let hx = self.hx.as_ref().map(|hx| (*hx.get_data()).to_ref());
+        let output = rnn_desc.rnn_fwd(
             self.x.get_as_ref(),
             hx,
-            cx,
             self.weight.get_as_ref(),
             self.is_training,
         );
         self.hy.upgrade().unwrap().swap_inner(output.hy);
-        self.cy.upgrade().unwrap().swap_inner(output.cy);
         self.y.upgrade().unwrap().swap_inner(output.y);
     }
 
-    #[expect(clippy::similar_names)]
     fn backward(&self) {
         assert!(
             self.is_training,
@@ -44,39 +40,25 @@ impl<T: Num> Function<T, Nvidia> for CudnnLSTM<T> {
 
         let mut rnn_desc = self.rnn_desc.borrow_mut();
 
-        let x_shape = self.x.get_shape();
+        let x = self.x.get_as_ref();
         let y = self.y.upgrade().unwrap().get_as_ref();
         let dy = self.y.upgrade().unwrap().get_grad().unwrap().get_as_ref();
-        let hx = self.hx.as_ref().map(Variable::get_as_ref);
-        let cx = self.cx.as_ref().map(Variable::get_as_ref);
+        let hx = self.hx.as_ref().map(|hx| (*hx.get_data()).to_ref());
         let dhy = self
             .hy
             .upgrade()
-            .map(|dhy| dhy.get_grad().unwrap().get_as_ref());
-        let dcy = self
-            .cy
-            .upgrade()
-            .map(|dcy| dcy.get_grad().unwrap().get_as_ref());
+            .unwrap()
+            .get_grad()
+            .map(|dhy| dhy.get_as_ref());
         let weight = self.weight.get_data().to_ref();
-        let ddata = rnn_desc.lstm_bkwd_data(
-            x_shape,
-            y.clone(),
-            dy,
-            hx.clone(),
-            cx.clone(),
-            dhy,
-            dcy,
-            weight,
-        );
+        let ddata =
+            rnn_desc.rnn_bkwd_data(self.x.get_shape(), y.clone(), dy, hx.clone(), dhy, weight);
 
-        let dw = rnn_desc.lstm_bkwd_weights(self.x.get_as_ref(), hx.clone(), cx.clone(), y.clone());
+        let dw = rnn_desc.rnn_bkwd_weights(x, hx, y);
 
         self.x.set_grad(Variable::new(ddata.dx));
         if let Some(hx) = self.hx.as_ref() {
             hx.set_grad(Variable::new(ddata.dhx));
-        }
-        if let Some(cx) = self.cx.as_ref() {
-            cx.set_grad(Variable::new(ddata.dcx));
         }
 
         self.weight.set_grad(Variable::new(dw));
@@ -87,37 +69,30 @@ impl<T: Num> Function<T, Nvidia> for CudnnLSTM<T> {
         if let Some(hx) = &self.hx {
             inputs.push(hx.clone());
         }
-        if let Some(cx) = &self.cx {
-            inputs.push(cx.clone());
-        }
         inputs.push(self.weight.clone());
         inputs
     }
 }
 
-pub fn lstm_cudnn<T: Num>(
+pub fn cudnn_rnn_fwd<T: Num>(
     rnn_desc: Rc<RefCell<RNNDescriptor<T>>>,
     x: Variable<T, Nvidia>,
     hx: Option<Variable<T, Nvidia>>,
-    cx: Option<Variable<T, Nvidia>>,
     weight: Variable<T, Nvidia>,
     is_training: bool,
-) -> Variable<T, Nvidia> {
+) -> RNNOutput<T, Nvidia> {
     let num_layers = rnn_desc.borrow().get_num_layers();
     let hidden_size = rnn_desc.borrow().get_hidden_size();
     let x_shape = x.get_shape();
     let seq_len = x_shape[0];
     let batch_size = x_shape[1];
     let hy = alloc([num_layers, batch_size, hidden_size]);
-    let cy = alloc([num_layers, batch_size, hidden_size]);
     let y = alloc([seq_len, batch_size, hidden_size]);
-    let layer = CudnnLSTM {
+    let layer = CudnnRNN {
         rnn_desc,
         x,
         hx,
-        cx,
         hy: hy.clone().downgrade(),
-        cy: cy.clone().downgrade(),
         y: y.clone().downgrade(),
         weight,
         is_training,
@@ -127,22 +102,20 @@ pub fn lstm_cudnn<T: Num>(
 
     let layer = Rc::new(RefCell::new(Box::new(layer) as Box<dyn Function<T, Nvidia>>));
     hy.set_creator(layer.clone());
-    cy.set_creator(layer.clone());
     y.set_creator(layer);
 
-    // LSTMOutput { y, hy, cy }
-    y
+    RNNOutput { y, hy }
 }
 
 #[cfg(test)]
-mod lstm_test {
+mod rnn_test {
     use std::{cell::RefCell, rc::Rc};
 
     use zenu_matrix::{
         device::{cpu::Cpu, nvidia::Nvidia},
         dim::DimDyn,
         matrix::{Matrix, Owned},
-        nn::rnn::{RNNDescriptor, RNNWeightsMat},
+        nn::rnn::{RNNDescriptor, RNNWeights},
     };
 
     use zenu_test::{
@@ -151,13 +124,13 @@ mod lstm_test {
 
     use crate::Variable;
 
-    use super::lstm_cudnn;
+    use super::cudnn_rnn_fwd;
 
     #[cfg(feature = "nvidia")]
     #[test]
-    fn lstm() {
+    fn rnn() {
         let matrix_map =
-            read_test_case_from_json_val!("../test_data_json/lstm_fwd_bkwd_small.json");
+            read_test_case_from_json_val!("../test_data_json/rnn_fwd_bkwd_single_seq_len_1.json");
 
         let mut weights = Vec::new();
         let input_weight: Matrix<Owned<f32>, DimDyn, Cpu> =
@@ -166,7 +139,7 @@ mod lstm_test {
         let input_bias = matrix_map.get("rnn.bias_ih_l0").unwrap().clone();
         let hidden_bias = matrix_map.get("rnn.bias_hh_l0").unwrap().clone();
 
-        let rnn_weights = RNNWeightsMat::new(input_weight, hidden_weight, input_bias, hidden_bias);
+        let rnn_weights = RNNWeights::new(input_weight, hidden_weight, input_bias, hidden_bias);
         weights.push(rnn_weights);
 
         let input = matrix_map.get("input").unwrap().clone();
@@ -176,7 +149,7 @@ mod lstm_test {
         let batch_size = input.shape()[1];
 
         let rnn_desc =
-            RNNDescriptor::<f32>::lstm(false, 0.0, input_size, hidden_size, 1, batch_size);
+            RNNDescriptor::<f32>::new_rnn_relu(false, 0.0, input_size, hidden_size, 1, batch_size);
 
         let input = matrix_map.get("input").unwrap().clone();
         let input = Variable::new(input.to::<Nvidia>());
@@ -191,20 +164,13 @@ mod lstm_test {
 
         let rnn_desc = Rc::new(RefCell::new(rnn_desc));
 
-        let result = lstm_cudnn(
-            rnn_desc.clone(),
-            input.clone(),
-            None,
-            None,
-            weight.clone(),
-            true,
-        );
+        let result = cudnn_rnn_fwd(rnn_desc.clone(), input.clone(), None, weight.clone(), true);
 
-        result.backward();
+        result.y.backward();
 
         let expected_output = matrix_map.get("output").unwrap();
 
-        assert_val_eq!(result, expected_output.clone().to::<Nvidia>(), 1e-5);
+        assert_val_eq!(result.y, expected_output.clone().to::<Nvidia>(), 1e-5);
         assert_val_eq_grad!(
             input.clone(),
             matrix_map.get("input_grad").unwrap().clone().to::<Nvidia>(),
