+++ zenu-layer/src/layers/mod.rs
@@ -2,6 +2,5 @@ pub mod batch_norm_2d;
 pub mod conv2d;
 pub mod dropout;
 pub mod linear;
-pub mod lstm;
 pub mod max_pool_2d;
 pub mod rnn;
