+++ zenu-layer/src/layers/rnn/mod.rs
@@ -1,0 +1,5 @@
+mod lstm;
+mod rnn;
+
+pub use lstm::*;
+pub use rnn::*;
